Applying review feedback - merged into single script, added option 'Pulse on select'
