--- Assets/MRTK/SDK/Experimental/PulseShader/Scripts/SurfacePulseOnSelect.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8c9e78ea1f93a324586fd98000ef5f56
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
